refactor(test): share node http test server helper
